Delete product and some fixes

+++ lib/models/section.dart
@@ -77,7 +77,7 @@ class Section extends ChangeNotifier {
     }
 
     for(final original in originalItems) {
-      if(!items.contains(original)){
+      if(!items.contains(original) && (original.image as String).contains('firebase')){
         try {
           final ref = await storage.getReferenceFromUrl(
               original.image as String
@@ -100,14 +100,16 @@ class Section extends ChangeNotifier {
     await firestoreRef.delete();
 
     for(final item in items) {
-      try {
-        final ref = await storage.getReferenceFromUrl(
-            item.image as String
-        );
+      if((item.image as String).contains('firebase')) {
+        try {
+          final ref = await storage.getReferenceFromUrl(
+              item.image as String
+          );
 
-        await ref.delete();
-        // ignore: empty_catches
-      } catch (e) {}
+          await ref.delete();
+          // ignore: empty_catches
+        } catch (e) {}
+      }
     }
   }
